Changed all to 64 bits (long int mode) now, the example with a : c = 11 + d - 7 works

+++ trunk/my-gcc/tests/executor.c
@@ -1,5 +1,5 @@
 #include "test.h"
 
-int main(int argc, char ** argv){
-  printf("False main result %d\n",falsemain());
+long int main(int argc, char ** argv){
+  printf("False main result %ld\n",falsemain());
 }
